Make LLIntAssembly.h get regenerated after changing architecture target
The ${LLIntOutput} file is architecture-specific but resides in a common place. When we build for Simulator,
it is created for the x86/x86_64 processor architectures. Afterwards all device builds will start failing
because they need arm7/arm64 architectures but the file will be considered up-to-date and won't be regenerated.
As a workaround we're creating an additional dependency of a stamp file per architecture. Whenever a specific
architecture is generated, all other ones will be invalidated by touching the corresponding dependency files.