feat(runtime): Allow JS objects to dynamically define properties when assigning to them
So far `JSObject::putInlineSlow` checked if the object already has the property defined in its
structure without giving it a chance to dynamically provide it. In order to stop NativeScript
prototype objects to have to materialize all of their properties in advance,
we need to call their `getOwnPropertySlot` function here.

Swallow any JS exceptions that may have been generated. When overwriting global inline
functions in `GlobalObject` a JS exception stating that the symbol is not available at runtime
is being thrown.

As an optimization, omit calling `JSFunction::getOwnProperty` and `JSObject::getOwnProperty`
if they haven't been overridden.